requirements.txt
git+https://github.com/nashspence/home-index.git
debugpy
easyocr
numpy
pillow
python-magic
rawpy
wand
opencv-python